requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-allauth==0.63.3
django-cors-headers==4.3.1
dj-rest-auth==6.0.0
djangorestframework-simplejwt==5.3.0
cryptography==42.0.5

# Gemini AI
google-generativeai==0.3.2

# Base de datos (opcional para PostgreSQL)
# psycopg2-binary==2.9.9

# Utilidades
python-decouple==3.8
Pillow==10.0.0